Control commands don't send the com-port-option and the command code, this is now done in BlockingCommandSender.

+++ org.xidobi.rfc2217.test/src/testtools/ByteBuffer.java
@@ -42,7 +42,7 @@ public class ByteBuffer {
 
 	public ByteBuffer putBytes(int... v) {
 		for (int i : v)
-			putInt(i);
+			putByte(i);
 
 		return this;
 	}
